server/scraper/requirements.txt
requests
beautifulsoup4
pymongo
google-generativeai
python-dotenv
